finished project

+++ src/save_file.py
@@ -9,19 +9,16 @@ class AbstractWorkFile(ABC):
     @abstractmethod
     def save_data(self, vacancy_list):
         """Абстрактный метод для сохранения вакансий"""
-
         pass
 
     @abstractmethod
     def get_data(self, keyword):
         """Абстрактный метод для получения данных"""
-
         pass
 
     @abstractmethod
     def del_data(self, keyword):
         """Абстрактный метод для удаления информации"""
-
         pass
 
 
@@ -33,24 +30,32 @@ class SaveFile(AbstractWorkFile):
 
     def save_data(self, vacancy_list):
         """Метод для сохранения вакансий"""
+        if not isinstance(vacancy_list, list):
+            vacancy_list = [vacancy_list]
 
         if os.path.exists(self.__file_name):
             with open(self.__file_name, 'r', encoding="UTF-8") as f:
                 data = json.load(f)
-            new_data = vacancy_list
-            data.append(new_data)
+                if not isinstance(data, list):
+                    data = [data]
+            data.extend(vacancy_list)
+
+            with open(self.__file_name, 'w', encoding="UTF-8") as f:
+                json.dump(data, f, ensure_ascii=False, indent=4)
 
         else:
             with open(self.__file_name, 'w', encoding="UTF-8") as f:
-                data = vacancy_list
-                json.dump(data, f)
-
+                json.dump(vacancy_list, f, ensure_ascii=False, indent=4)
 
     def get_data(self, keyword):
         """Метод для получения данных по ключевому слову"""
+        if not os.path.exists(self.__file_name):
+            return []
 
         with open(self.__file_name, 'r', encoding='UTF-8') as f:
             data = json.load(f)
+            if not isinstance(data, list):
+                data = [data]
         result = []
         for item in data:
             if (keyword.lower() in item.get("name", "").lower() or
@@ -60,17 +65,19 @@ class SaveFile(AbstractWorkFile):
 
     def del_data(self, keyword):
         """Метод для удаления информации"""
+        if not os.path.exists(self.__file_name):
+            return
 
         with open(self.__file_name, 'r', encoding='UTF-8') as f:
-            data = json.load(f)
+                data = json.load(f)
+                if not isinstance(data, list):
+                    data = [data]
+
         filtered_data = [
             item for item in data
-            if (item.get("name", "").lower() != keyword.lower() and
+            if (keyword.lower() not in item.get("name", "").lower() and
                 keyword.lower() not in item.get("responsibility", "").lower())
         ]
+
         with open(self.__file_name, 'w', encoding='UTF-8') as f:
             json.dump(filtered_data, f, ensure_ascii=False, indent=4)
-
-
-
-
